requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
typing_extensions==4.15.0

# Core OCR and image processing
pytesseract
opencv-python
Pillow

# PDF processing
pdfplumber
pdf2image

# Data handling and output
pandas

# Additional dependencies
numpy